+++ 02_react/package.json
@@ -11,7 +11,8 @@
   },
   "dependencies": {
     "react": "^19.1.1",
-    "react-dom": "^19.1.1"
+    "react-dom": "^19.1.1",
+    "react-router": "^7.8.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.33.0",
