+++ 02_react/src/assets/icons/icons.ts
@@ -1,0 +1,44 @@
+export { default as Ifilled_bookmark } from './icon/filled/bookmark.svg';
+export { default as Ifilled_charge } from './icon/filled/charge.svg';
+export { default as Ifilled_check } from './icon/filled/check.svg';
+export { default as Ifilled_down_arrow } from './icon/filled/down_arrow.svg';
+export { default as Ifilled_mypage } from './icon/filled/mypage.svg';
+export { default as Ifilled_star } from './icon/filled/star.svg';
+export { default as Ifilled_store } from './icon/filled/store.svg';
+export { default as Ifilled_triptalk } from './icon/filled/triptalk.svg';
+export { default as Ifilled_up_arrow } from './icon/filled/up_arrow.svg';
+export { default as Ifilled_visibility } from './icon/filled/visibility.svg';
+
+export { default as Ioutline__spa } from './icon/outline/_spa.svg';
+export { default as Ioutline_add } from './icon/outline/add.svg';
+export { default as Ioutline_apartment } from './icon/outline/apartment.svg';
+export { default as Ioutline_bad } from './icon/outline/bad.svg';
+export { default as Ioutline_bookmark } from './icon/outline/bookmark.svg';
+export { default as Ioutline_calendar } from './icon/outline/calendar.svg';
+export { default as Ioutline_camp } from './icon/outline/camp.svg';
+export { default as Ioutline_chat } from './icon/outline/chat.svg';
+export { default as Ioutline_close } from './icon/outline/close.svg';
+export { default as Ioutline_edit } from './icon/outline/edit.svg';
+export { default as Ioutline_fire } from './icon/outline/fire.svg';
+export { default as Ioutline_good } from './icon/outline/good.svg';
+export { default as Ioutline_hotel } from './icon/outline/hotel.svg';
+export { default as Ioutline_house_on_the_sea } from './icon/outline/house_on_the_sea.svg';
+export { default as Ioutline_left_arrow } from './icon/outline/left_arrow.svg';
+export { default as Ioutline_link } from './icon/outline/link.svg';
+export { default as Ioutline_location } from './icon/outline/location.svg';
+export { default as Ioutline_login } from './icon/outline/login.svg';
+export { default as Ioutline_logout } from './icon/outline/logout.svg';
+export { default as Ioutline_menu } from './icon/outline/menu.svg';
+export { default as Ioutline_person } from './icon/outline/person.svg';
+export { default as Ioutline_planterior } from './icon/outline/planterior.svg';
+export { default as Ioutline_point } from './icon/outline/point.svg';
+export { default as Iradio_enabled } from './icon/outline/radio-enabled.svg';
+export { default as Iradio_selected } from './icon/outline/radio-selected.svg';
+export { default as Ioutline_reply } from './icon/outline/reply.svg';
+export { default as Ioutline_return } from './icon/outline/return.svg';
+export { default as Ioutline_right_arrow } from './icon/outline/right_arrow.svg';
+export { default as Ioutline_room_service } from './icon/outline/room service.svg';
+export { default as Ioutline_rwite } from './icon/outline/rwite.svg';
+export { default as Ioutline_search } from './icon/outline/search.svg';
+export { default as Ioutline_Single_person_accommodation } from './icon/outline/Single person accommodation.svg';
+export { default as Ioutline_uncheck } from './icon/outline/uncheck.svg';
