+++ 02_react/src/index.css
@@ -1,5 +1,5 @@
 :root {
-  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
+  font-family: 'Pretendard Variable';
   line-height: 1.5;
   font-weight: 400;
